refactor(ScoreCalculator) : 점수 계산 책임을 별도의 객체에 위임
