+++ src/main/java/chess/domain/Board.java
@@ -9,9 +9,6 @@ import java.util.stream.Stream;
 
 public class Board {
 
-    private static final int MINUS_TARGET_SIZE = 2;
-    private static final double MINUS_SCORE = 0.5;
-
     private final Map<Position, Piece> board;
 
     public Board(Map<Position, Piece> board) {
@@ -41,39 +38,6 @@ public class Board {
 
     public boolean isEmptySpace(Position position) {
         return !hasPiece(position);
-    }
-
-    public Map<Color, Double> calculateScore() { // TODO : 보드의 책임이 맞을까??
-        return Stream.of(
-                        Map.entry(Color.WHITE, calculateScore(Color.WHITE)),
-                        Map.entry(Color.BLACK, calculateScore(Color.BLACK)))
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-    }
-
-    private double calculateScore(Color color) {
-        return calculateScoreBeforeMinus(color) - calculateMinusScore(color);
-    }
-
-    private double calculateScoreBeforeMinus(Color color) {
-        return board.values().stream()
-                .filter(piece -> piece.isSameTeam(color))
-                .mapToDouble(Piece::getScore)
-                .sum();
-    }
-
-    private double calculateMinusScore(Color color) {
-        Map<Column, Long> pawnBoard = board.keySet().stream()
-                .filter(position -> board.get(position).isSameTeam(color))
-                .filter(position -> board.get(position).getPieceType().isPawn())
-                .collect(Collectors.groupingBy(Position::getColumn, Collectors.counting()));
-
-        long sameLinePawnCount = pawnBoard.keySet()
-                .stream()
-                .filter(column -> pawnBoard.get(column) >= MINUS_TARGET_SIZE)
-                .map(pawnBoard::get)
-                .reduce(0L, Long::sum);
-
-        return MINUS_SCORE * sameLinePawnCount;
     }
 
     public Map<Position, Piece> getBoard() {
